pyproject.toml
[tool.poetry]
name = "crm"
version = "0.1.0"
description = "chatbot"
authors = ["Zeta Labs"]
readme = "README.md"
license = "MIT"

# Including only the core package and specific files
packages = [
    { include = "crm" }  # Only include the main code
]

# exclude the test directory or any unnecessary files
exclude = [
    "tests/*",  # Exclude standalone tests from the package distribution
]

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
qdrant-client = "^1.14.3"

# Core RAG components
langchain-community = "^0.3.24"

# Web API
fastapi = "^0.115.12"
uvicorn = "^0.34.2"

# NLP utils
langdetect = "^1.0.9"
docx2txt = "^0.9"
unstructured = "^0.17.2"           # For document parsing, keep only if needed

# Databases / Caches
pymongo = "^4.13.0"
redis = "^6.2.0"

# Multimedia
langchain-ollama = ">=0.3.2,<0.4.0"
langchain-openai = ">=0.1.0"       # For OpenAI integration
openai = ">=1.0.0"                 # OpenAI Python SDK
tiktoken = "^0.8.0"                # Fast tokenizer for OpenAI models
pika = "^1.3.2"
langcodes = "^3.5.0"
pyjwt = "^2.10.1"
pdf2image = "^1.17.0"
python-docx = "^1.2.0"
weasyprint = "^66.0"
pillow = "^11.3.0"
python-multipart = "^0.0.20"

[tool.poetry.group.dev.dependencies]
# For development, debugging, notebooks
pika = ">=1.3.2,<2.0.0"
python-dotenv = ">=1.1.0,<2.0.0"
langchain = ">=0.3.25,<0.4.0"
pypdf = ">=5.4.0,<6.0.0"
notebook = ">=7.4.2"
numpy = ">=1.26"
pandas = ">=2.0.0"
requests = ">=2.28.0"
